📦️ refactor: Extract GitHub logic into @liam-hq/github package
