+++ frontend/apps/app/features/projects/pages/ProjectNewPage/InstallationSelector/InstallationSelector.tsx
@@ -9,8 +9,8 @@ import {
 } from '@/components'
 import { addProject } from '@/features/projects/actions'
 import { createClient } from '@/libs/db/client'
-import { getRepositoriesByInstallationId } from '@/libs/github/api.browser'
-import type { Installation, Repository } from '@/libs/github/types'
+import { getRepositoriesByInstallationId } from '@liam-hq/github'
+import type { Installation, Repository } from '@liam-hq/github'
 import { type FC, useCallback, useEffect, useState } from 'react'
 import { P, match } from 'ts-pattern'
 import { RepositoryItem } from '../RepositoryItem'
