--- frontend/apps/app/libs/github/types.ts
@@ -1,4 +1,0 @@
-import type { components } from '@octokit/openapi-types'
-
-export type Installation = components['schemas']['installation']
-export type Repository = components['schemas']['repository']
