+++ frontend/packages/github/.gitignore
@@ -1,0 +1,6 @@
+# env
+.env
+.env.local
+
+# build
+dist
