+++ frontend/packages/github/biome.jsonc
@@ -1,0 +1,3 @@
+{
+  "extends": ["../../packages/configs/biome.jsonc"]
+}
