+++ frontend/packages/github/package.json
@@ -1,0 +1,33 @@
+{
+  "name": "@liam-hq/github",
+  "private": true,
+  "version": "0.1.0",
+  "main": "./dist/index.js",
+  "exports": {
+    ".": "./dist/index.js"
+  },
+  "dependencies": {
+    "@octokit/auth-app": "6.1.3",
+    "@octokit/openapi-types": "24.0.0",
+    "@octokit/rest": "20.0.2",
+    "@supabase/supabase-js": "2.49.1"
+  },
+  "devDependencies": {
+    "@biomejs/biome": "1.9.3",
+    "@liam-hq/configs": "workspace:*",
+    "@types/node": "22.9.0",
+    "typescript": "5",
+    "vitest": "3.0.8"
+  },
+  "scripts": {
+    "build": "tsc",
+    "dev": "tsc --watch",
+    "fmt": "pnpm run '/^fmt:.*/'",
+    "fmt:biome": "biome check --write --unsafe .",
+    "lint": "pnpm run '/^lint:.*/'",
+    "lint:biome": "biome check .",
+    "lint:tsc": "tsc --noEmit",
+    "test": "vitest --watch=false"
+  },
+  "types": "./dist/index.d.ts"
+}
